requirements.txt
Flask==0.10.1
Flask-Script==0.5.3
Jinja2==2.7
MarkupSafe==0.18
Werkzeug==0.9.2
distribute==0.6.34
itsdangerous==0.22
wsgiref==0.1.2
google-api-python-client==1.1
sendgrid==0.1.4
WTForms==1.0.4
Flask-SQLAlchemy==0.16
MySQL-python==1.2.4